+++ libs/JSystem/include/JSystem/JKernel/JKRHeap.hpp
@@ -2,6 +2,7 @@
 #define JKR_HEAP_H
 
 #include <dolphin/types.h>
+#include <stdint.h>
 #include <dolphin/os/OSMutex.h>
 #include <JSystem/JKernel/JKRDisposer.hpp>
 #include <new>
@@ -71,7 +72,7 @@ public:
 	void* getMaxFreeBlock();
 	u32 getMaxAllocatableSize(int alignment);
 	JKRHeap* find(void*) const;
-	void dispose_subroutine(u32 begin, u32 end);
+	void dispose_subroutine(uintptr_t begin, uintptr_t end);
 	bool dispose(void*, u32);
 	void dispose(void*, void*);
 	void dispose();
