+++ libs/JSystem/src/J3D/J3DGraphLoader/J3DModelLoader.cpp
@@ -240,7 +240,7 @@ void J3DModelLoader::readVertex(const J3DVertexBlock* i_block)
 		vertex_data.mNrmNum = 0;
 	} else if (nrm_end != NULL) {
 		vertex_data.mNrmNum
-		    = ((u32)nrm_end - (u32)vertex_data.mVtxNormArray) / 12 + 1;
+		    = ((uintptr_t)nrm_end - (uintptr_t)vertex_data.mVtxNormArray) / 12 + 1;
 	} else {
 		vertex_data.mNrmNum
 		    = (i_block->mSize - (u32)i_block->mpVtxNrmArray) / 12 + 1;
@@ -257,7 +257,7 @@ void J3DModelLoader::readVertex(const J3DVertexBlock* i_block)
 		vertex_data.mColNum = 0;
 	} else if (color0_end != NULL) {
 		vertex_data.mColNum
-		    = ((u32)color0_end - (u32)vertex_data.mVtxColorArray[0]) / 4 + 1;
+		    = ((uintptr_t)color0_end - (uintptr_t)vertex_data.mVtxColorArray[0]) / 4 + 1;
 	} else {
 		vertex_data.mColNum
 		    = (i_block->mSize - (u32)i_block->mpVtxColorArray[0]) / 4 + 1;
@@ -331,7 +331,7 @@ void J3DModelLoader_v26::readMaterial(const J3DMaterialBlock* i_block,
 	if (i_flags & J3DMLF_UseUniqueMaterials) {
 		for (u16 i = 0; i < mpModelData->unk34; i++) {
 			factory.create(&mpModelData->unk38[i], i, i_flags);
-			mpModelData->unk38[i].unk18 = (u32)&mpModelData->unk38[i] >> 4;
+			mpModelData->unk38[i].unk18 = (uintptr_t)&mpModelData->unk38[i] >> 4;
 		}
 	}
 	for (u16 i = 0; i < mpModelData->mMaterialNum; i++) {
@@ -373,7 +373,7 @@ void J3DModelLoader_v21::readMaterial_v21(const J3DMaterialBlock_v21* i_block,
 	if (i_flags & J3DMLF_UseUniqueMaterials) {
 		for (u16 i = 0; i < mpModelData->unk34; i++) {
 			factory.create(&mpModelData->unk38[i], i, i_flags);
-			mpModelData->unk38[i].unk18 = (u32)&mpModelData->unk38[i] >> 4;
+			mpModelData->unk38[i].unk18 = (uintptr_t)&mpModelData->unk38[i] >> 4;
 		}
 	}
 	for (u16 i = 0; i < mpModelData->mMaterialNum; i++) {
@@ -451,7 +451,7 @@ void J3DModelLoader_v26::readMaterialTable(const J3DMaterialBlock* i_block,
 	}
 	for (u16 i = 0; i < mpMaterialTable->mMaterialNum; i++) {
 		mpMaterialTable->mMaterials[i]->unk18
-		    = (u32)mpMaterialTable->mMaterials + factory.getMaterialID(i);
+		    = (uintptr_t)mpMaterialTable->mMaterials + factory.getMaterialID(i);
 	}
 }
 
@@ -473,7 +473,7 @@ void J3DModelLoader_v21::readMaterialTable_v21(
 	}
 	for (u16 i = 0; i < mpMaterialTable->mMaterialNum; i++) {
 		mpMaterialTable->mMaterials[i]->unk18
-		    = ((u32)mpMaterialTable->mMaterials) + factory.getMaterialID(i);
+		    = ((uintptr_t)mpMaterialTable->mMaterials) + factory.getMaterialID(i);
 	}
 }
