+++ libs/JSystem/src/JKernel/JKRDvdAramRipper.cpp
@@ -64,7 +64,7 @@ bool JKRDvdAramRipper::errorRetry = true;
 
 JKRADCommand* JKRDvdAramRipper::loadToAram_Async(char* path, u32 address,
                                                  JKRExpandSwitch expandSwitch,
-                                                 void (*callback)(u32),
+                                                 void (*callback)(uintptr_t),
                                                  u32 param_4, u32 param_5)
 {
 	// The linker takes the two JGadget::TPointer<JKRDvdFile> destructors from
@@ -76,7 +76,7 @@ JKRADCommand* JKRDvdAramRipper::loadToAram_Async(char* path, u32 address,
 
 JKRADCommand* JKRDvdAramRipper::loadToAram_Async(s32 entryNumber, u32 address,
                                                  JKRExpandSwitch expandSwitch,
-                                                 void (*callback)(u32),
+                                                 void (*callback)(uintptr_t),
                                                  u32 param_4, u32 param_5)
 {
 	JUT_ASSERT_F(false, "UNIMPLEMENTED");
@@ -86,7 +86,7 @@ JKRADCommand* JKRDvdAramRipper::loadToAram_Async(s32 entryNumber, u32 address,
 JKRADCommand* JKRDvdAramRipper::loadToAram_Async(JKRDvdFile* dvdFile,
                                                  u32 address,
                                                  JKRExpandSwitch expandSwitch,
-                                                 void (*callback)(u32),
+                                                 void (*callback)(uintptr_t),
                                                  u32 param_4, u32 param_5)
 {
 	JKRADCommand* command  = new (JKRGetSystemHeap(), -4) JKRADCommand();
@@ -132,7 +132,7 @@ JKRADCommand* JKRDvdAramRipper::callCommand_Async(JKRADCommand* command)
 
 		if (command->mExpandSwitch == 1) {
 			u8 buffer[0x40];
-			u8* bufPtr = (u8*)ALIGN_NEXT((u32)&buffer, 0x20);
+			u8* bufPtr = (u8*)ALIGN_NEXT((uintptr_t)&buffer, 0x20);
 			while (true) {
 				if (DVDReadPrio(dvdFile->getFileInfo(), bufPtr, 0x20, 0, 2)
 				    >= 0) {
@@ -209,7 +209,7 @@ JKRADCommand* JKRDvdAramRipper::callCommand_Async(JKRADCommand* command)
 		if (!command->mCallback) {
 			sDvdAramAsyncList.append(&command->mLink);
 		} else {
-			command->mCallback((u32)command);
+			command->mCallback((uintptr_t)command);
 		}
 	}
 
@@ -473,7 +473,7 @@ static u32 dmaBufferFlush(u32 param_1)
 		return 0;
 	}
 	u32 size = ALIGN_NEXT(dmaCurrent - dmaBuf, 0x20);
-	JKRAramPiece::orderSync(0, (u32)dmaBuf, param_1, size, nullptr);
+	JKRAramPiece::orderSync(0, (uintptr_t)dmaBuf, param_1, size, nullptr);
 	dmaCurrent = dmaBuf;
 	return size;
 }
